replace existing page with tabbed routed page
