+++ src/components/pages/article/article-page.module.scss
@@ -1,6 +1,7 @@
 @use "../../../sass/settings";
 @use "../../../sass/layout";
 @use "../../../sass/mixins";
+@use "../../../sass/typography";
 
 .article-page {
   @include settings.container-padding();
@@ -24,5 +25,82 @@
     flex: 0 0 auto;
     max-width: 180px;
     width: 100%;
+  }
+}
+
+.tabbed-navigation__tabs {
+  @include mixins.margin(0 -24 36);
+  @include mixins.padding(0, "left");
+
+  align-self: flex-start;
+  background-color: #fff;
+  border-bottom: settings.$border;
+  border-top: settings.$border;
+  display: flex;
+  justify-content: center;
+  list-style: none;
+  position: sticky;
+  top: 0;
+
+  @media only screen and (min-width: settings.$breakpoint-small) {
+    @include mixins.margin(0 -48 36);
+  }
+
+  @media only screen and (min-width: settings.$breakpoint-medium) {
+    @include mixins.margin(0 0 36);
+  }
+
+  @media only screen and (min-width: settings.$breakpoint-large) {
+    border-top: none;
+  }
+}
+
+.tabbed-navigation__tab-label {
+  @include typography.body();
+
+  color: settings.$color-text-inactive;
+  cursor: pointer;
+  display: inline;
+  font-family: settings.$font-family-primary;
+
+  a {
+    @include mixins.margin(0 6);
+    @include mixins.padding(18 6);
+    color: settings.$color-text-inactive;
+    text-decoration: none;
+    display: block
+  }
+
+  &.tabbed-navigation__tab-label--active {
+    @include mixins.padding(0, "bottom");
+
+    border-bottom: 4px solid settings.$color-primary;
+    color: settings.$color-text;
+    font-weight: 600;
+
+    a {
+      color: settings.$color-text;
+    }
+  }
+
+  @media only screen and (max-width: settings.$breakpoint-small) {
+    &:nth-child(2) {
+      overflow: hidden;
+      white-space: nowrap;
+      width: 52px;
+    }
+  }
+
+  @media only screen and (min-width: settings.$breakpoint-small) {
+    a {
+      @include mixins.margin(0);
+      @include mixins.padding(24 18);
+    }
+  }
+}
+
+.tabbed-navigation__content {
+  @media only all and (min-width: settings.$breakpoint-large) {
+    display: flex;
   }
 }
