+++ src/components/pages/article/article-page.tsx
@@ -1,20 +1,12 @@
-import { useState } from 'react';
-import { ArticleContent } from '../../atoms/article-content/article-content';
-import { Heading } from '../../atoms/heading/heading';
-import { JumpToMenu } from '../../atoms/jump-to-menu/jump-to-menu';
+import { ReactElement, useState } from 'react';
+import Link from 'next/link';
 import { ArticleStatus } from '../../molecules/article-status/article-status';
 import { ContentHeader } from '../../molecules/content-header/content-header';
 import { SiteHeader } from '../../molecules/site-header/site-header';
-import { Tab, TabbedNavigation } from '../../molecules/tabbed-navigation';
 import { Timeline, TimelineEvent } from '../../molecules/timeline/timeline';
-import { Content } from '../../../types/content';
 import styles from './article-page.module.scss';
-import { EditorsAndReviewers } from '../../atoms/editors-and-reviewers/editors-and-reviewers';
-import { ReviewContent } from '../../atoms/review-content/review-content';
-import { Abstract } from '../../atoms/abstract/abstract';
-import { ReferenceList } from '../../atoms/reference-list/reference-list';
-import { AuthorList } from '../../molecules/author-list/author-list';
-import { MetaData, PeerReview } from '../../../types';
+import { MetaData } from '../../../types';
+import { ArticleFiguresTab, ArticleFullTextTab, ArticleReviewsTab } from './tabs';
 
 export type ArticleStatusProps = {
   timeline: TimelineEvent[],
@@ -22,28 +14,35 @@ export type ArticleStatusProps = {
   status: string,
 };
 
-const getFigures = (content: Content): Content => {
-  if (typeof content === 'undefined') {
-    return '';
-  }
-  if (typeof content === 'string') {
-    return content;
-  }
-
-  if (Array.isArray(content)) {
-    return content.map((part) => getFigures(part));
-  }
-  switch (content.type) {
-    case 'Figure':
-      return content;
-    default:
-      return '';
-  }
+export type Tab = {
+  id: string,
+  element: ReactElement,
 };
 
-export const ArticlePage = (props: { metaData: MetaData, content: Content, status: ArticleStatusProps, peerReview: PeerReview }): JSX.Element => {
-  const [activeTab, setActiveTab] = useState<number>(0);
+export type ArticlePageProps = {
+  metaData: MetaData,
+  status: ArticleStatusProps,
+  children: ReactElement<typeof ArticleFullTextTab | typeof ArticleFiguresTab | typeof ArticleReviewsTab>,
+  activeTab: 'fulltext' | 'figures' | 'reviews',
+  tabs?: Tab[],
+};
 
+export const ArticlePage = (props: ArticlePageProps): JSX.Element => {
+  const [activeTab, setActiveTab] = useState<string>(props.activeTab);
+  const tabs = props.tabs ?? [
+    {
+      id: 'fulltext',
+      element: <Link href={`/reviewed-preprints/${props.metaData.msid}`}>Full text</Link>,
+    },
+    {
+      id: 'figures',
+      element: <Link href={`/reviewed-preprints/${props.metaData.msid}/figures`}>Figures and data</Link>,
+    },
+    {
+      id: 'reviews',
+      element: <Link href={`/reviewed-preprints/${props.metaData.msid}/reviews`}>Peer review</Link>,
+    },
+  ];
   return (
     <div className={`${styles['grid-container']} ${styles['article-page']}`}>
       <div className={styles['grid-header']}>
@@ -62,40 +61,16 @@ export const ArticlePage = (props: { metaData: MetaData, content: Content, statu
         <Timeline events={props.status.timeline}/>
       </aside>
       <main className={styles['primary-section']}>
-        <TabbedNavigation activeTab={activeTab} setActiveTab={setActiveTab}>
-          <Tab label="Full text">
-            <JumpToMenu headings={[
-              { id: 'abstract', text: 'Abstract' },
-              { id: 'assessment', text: 'eLife assessment' },
-              ...props.metaData.headings,
-              { id: 'references', text: 'References' },
-              { id: 'author-list', text: 'Author Information' },
-            ]} />
-            <div className={styles['article-body-container']}>
-              <Abstract content={props.metaData.abstract} />
-              <ReviewContent content={props.peerReview.evaluationSummary.text} isAssessment={true} setActiveTab={setActiveTab}/>
-              <ArticleContent content={props.content} />
-              <ReferenceList references={props.metaData.references} />
-              <AuthorList authors={props.metaData.authors}/>
-            </div>
-          </Tab>
-          <Tab label="Figures and data">
-            <div className={styles['menu-spacer']}/>
-            <div className={styles['article-body-container']}>
-              <Heading id="figures" headingLevel={2} content="Figures and data" />
-              <ArticleContent content={getFigures(props.content)} />
-            </div>
-          </Tab>
-          <Tab label="Peer review">
-            <div className={styles['menu-spacer']}/>
-            <div className={styles['article-body-container']}>
-              <EditorsAndReviewers participants={props.peerReview.evaluationSummary.participants} />
-              {props.peerReview.reviews.map((review, index) => (
-                <ReviewContent key={index} id={`peer-review-${index}`} content={review.text} />
-              ))}
-            </div>
-          </Tab>
-        </TabbedNavigation>
+      <div className={styles['tabbed-navigation']}>
+        <ul className={styles['tabbed-navigation__tabs']}>
+          {tabs.map((tab, index) => (
+            <li key={index} className={`${styles['tabbed-navigation__tab-label']}${activeTab === tab.id ? ` ${styles['tabbed-navigation__tab-label--active']}` : ''}`} onClick={() => setActiveTab(tab.id)}>
+              {tab.element}
+            </li>
+          ))}
+        </ul>
+      </div>
+      {props.children}
       </main>
     </div>
   );
