+++ src/components/pages/article/tabs/reviews-tab.tsx
@@ -1,4 +1,4 @@
-import styles from '../article-page-layout.module.scss';
+import styles from '../article-page.module.scss';
 import { EditorsAndReviewers } from '../../../atoms/editors-and-reviewers/editors-and-reviewers';
 import { ReviewContent } from '../../../atoms/review-content/review-content';
 import { PeerReview } from '../../../../types';
